Address PR review findings
- Grind picker (critical): an RPM-capable grinder with no RPM set opened the
  RPM wheel on its lowest seed (~750) and Done committed it. Seed the wheel at
  the mid-range default instead; Done commits whatever the wheels show.
- rpmStepSize is now grinder-model-wide (grinderWideRpmStep), matching the
  widget's grindRpmStepForGrinder — same fix already applied to the grind step.
- Log SQL exec() failures in grinderWideStep and the queryGrinderContext rpm
  branch (every other query in the file does; field AIs read these logs).
- GrinderContext rpmMin/rpmMax and the comparison-model rpm are now int, so MCP
  JSON emits 1200 not 1200.0.
- Fix stale comments: GrindPickerDialog header (NoAutoClose, not tap-outside),
  AI_ADVISOR.md "modal gap" -> smallest-repeated, shotprojection "above"->"below".
- Tests: deriveGrindStep 0.05 floor, no-repeat fallback, grinder-wide-across-beans;
  buildShotChangeDiff rpm integer diff + unchanged/zero skip.

+++ src/history/shothistory_types.h
@@ -235,11 +235,12 @@ struct GrinderContext {
     // 0 when fewer than 2 distinct numeric settings are available.
     double stepSize = 0;
     // RPM axis (variable-RPM grinders): the second half of the dial-in. Empty /
-    // 0 when the grinder has no recorded RPM history. rpmStepSize uses the same
-    // noise-filtered estimator as stepSize.
+    // 0 when the grinder has no recorded RPM history. RPM is an integer domain,
+    // so min/max are int; rpmStepSize stays double (reuses the same estimator as
+    // stepSize) and is grinder-model-wide, not bean-scoped.
     QList<int> rpmsObserved;
-    double rpmMin = 0;
-    double rpmMax = 0;
+    int rpmMin = 0;
+    int rpmMax = 0;
     double rpmStepSize = 0;
 };
